FE: Add rate setup for users

+++ pcforge-backend/src/main/java/com/kubacki/dawid/PCForge/models/setups/RatingSetup.java
@@ -22,5 +22,5 @@ public class RatingSetup {
     private ComputerSetup computerSetup;
 
     @Column(nullable = false)
-    private float rating;
+    private Double rating;
 }
